[Rohit] Displays error when user tries to delete a image which he/she has not created

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -149,7 +149,12 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttributes) {
+        Image image = imageService.getImage(imageId);
+        if(!isImgOwnerLoggedIn(image, session)){
+            redirectAttributes.addFlashAttribute("deleteError", true);
+            return "redirect:/images/" + image.getId() + '/' + image.getTitle();
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
